refactor: update import paths for ContentState to maintain consistent directory structure
