+++ src/lib/VideoState.svelte.ts
@@ -1,6 +1,6 @@
 import { getContext, setContext } from 'svelte';
 import type { ClipObject } from './types';
-import type { ContentState } from './ContentState.svelte';
+import type { ContentState } from './content/ContentState.svelte';
 
 const VIDEO_KEY = Symbol('video');
